Added support for the nicira extension sample action
Allow ofctl_rest to correctly parse a json representation of
the sample action.

In ofctl_rest, ast.literal_eval is used to parse json and turn it into a
python dict.

ofctl_utils.py is responsible for serializing the python dict
and outputing an OpenFlow message. Currently, any experiminter action is
being serialized as an unknown experimenter action. Therefore, in order
to support the sample action, we serialize into the correct structure
when we encounter that subtype.